+++ examples/nextjs/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@bento-editor/example-nextjs",
-  "version": "0.1.0",
-  "private": true,
+  "version": "0.0.1",
   "scripts": {
     "dev": "next dev"
   },
