+++ packages/core/src/editor/index.tsx
@@ -1,15 +1,14 @@
-import React, { useCallback, useEffect, useMemo } from 'react';
+import React, { useEffect, useMemo } from 'react';
 import { DndProvider } from 'react-dnd'
 import { HTML5Backend } from 'react-dnd-html5-backend'
 import { BaseEditor, createEditor } from 'slate';
 import { ReactEditor, Slate, withReact } from 'slate-react';
 import { Config } from '../config';
-import { Editable, EditableProps } from '../editable';
-import { Modal, useModal } from '../portals/modal';
+import { Editable } from '../editable';
 import { ModalContainer } from '../portals/modal/container';
 import { PopoverContainer } from '../portals/popover/container';
 import { GlobalStateProvider, useConfigGlobalStateSet, useScreenGlobalStateSet } from '../store';
-import { TextToolbar } from '../textToolbar';
+import { Toolbar } from '../toolbar';
 import { debounce } from '../utils';
 import { /*themeClass,*/ styles } from './index.css';
 
@@ -29,21 +28,18 @@ declare module 'slate' {
 type SlateProps = React.ComponentProps<typeof Slate>;
 
 export type EditorProps = {
-  initialValue: SlateProps['value'];
   config: Config;
+  initialValue: SlateProps['value'];
+  onChange: NonNullable<SlateProps['onChange']>;
 };
-export const Editor: React.FC<EditorProps> = ({ initialValue, config }) => {
+export const Editor: React.FC<EditorProps> = ({ config, initialValue, onChange }) => {
   const editor = useMemo(() => {
     return withReact(createEditor());
   }, []);
 
-  const handleChange = useCallback<NonNullable<SlateProps['onChange']>>((value) => {
-    console.log('Slate[onChange]', value);
-  }, []);
-
   return (
     <>
-      <Slate editor={editor} value={initialValue} onChange={handleChange}>
+      <Slate editor={editor} value={initialValue} onChange={onChange}>
         <GlobalStateProvider>
           <DndProvider backend={HTML5Backend}>
             {/* Need to wrap with a react component to encapsulate all state related processes inside the RecoilRoot component. */}
@@ -59,38 +55,6 @@ type RootProps = {
   config: Config;
 };
 const Root: React.FC<RootProps> = ({ config }) => {
-  const renderElement = useCallback<EditableProps['renderElement']>((props) => {
-    const { elements } = config;
-    const element = elements.find((element) => {
-      return element.type === props.element.type;
-    })
-    if (element) {
-      return (
-        <element.Component {...props}>{props.children}</element.Component>
-      );
-    } else {
-      return (
-        <div {...props.attributes}>{props.children}</div>
-      );
-    }
-  }, [config]);
-
-  const renderLeaf = useCallback<EditableProps['renderLeaf']>((props) => {
-    const { texts } = config;
-    const text = texts.find((text) => {
-      return text.type === props.text.type;
-    })
-    if (text) {
-      return (
-        <text.Component {...props}>{props.children}</text.Component>
-      );
-    } else {
-      return (
-        <span {...props.attributes}>{props.children}</span>
-      );
-    }
-  }, []);
-
   const setConfig = useConfigGlobalStateSet();
   useEffect(() => {
     setConfig(config);
@@ -121,23 +85,16 @@ const Root: React.FC<RootProps> = ({ config }) => {
     };
   }, [setScreen]);
 
-  const modal = useModal();
-  const handleModalOpenerClick = useCallback(() => {
-    modal.open();
-  }, [modal]);
-
   return (
     <>
       <div className={styles.root}>
         <div className={styles.container}>
-          <Editable renderElement={renderElement} renderLeaf={renderLeaf} />
+          <Editable />
         </div>
         <ModalContainer />
         <PopoverContainer />
       </div>
-      <TextToolbar />
-      <button onClick={handleModalOpenerClick}>open odal</button>
-      <Modal {...modal.bind}>hello</Modal>
+      <Toolbar />
     </>
   );
 };
