+++ packages/core/src/element/index.ts
@@ -1,6 +1,1 @@
-import React from 'react';
-import { RenderElementProps } from 'slate-react';
-
 export * from './container';
-
-export type Element = React.FC<RenderElementProps>;
