+++ packages/core/src/index.ts
@@ -1,4 +1,3 @@
 export * from './config';
 export * from './editor';
 export * from './element';
-export * from './text';
