--- packages/core/src/text/index.tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import { RenderLeafProps } from 'slate-react';
-
-export type Text = React.FC<RenderLeafProps>;
