+++ packages/element-paragraph/src/index.ts
@@ -1,1 +1,10 @@
-export * from './element';
+import { Element } from '@bento-editor/core';
+import editable from './editable';
+import toolbox from './toolbox';
+
+const element: Element = {
+  type: 'paragraph',
+  editable,
+  toolbox,
+};
+export default element;
