+++ packages/element-paragraph/src/toolbox/index.tsx
@@ -1,0 +1,12 @@
+import { Element } from '@bento-editor/core';
+import React from 'react';
+
+const toolbox: Element['toolbox'] = {
+  Icon: () => {
+    return (
+      <div>P</div>
+    );
+  },
+
+};
+export default toolbox;
