+++ packages/text-format/src/editable/index.tsx
@@ -1,0 +1,12 @@
+import { Text } from '@bento-editor/core';
+import React from 'react';
+
+const editable: Text['editable'] = {
+  Component: (props) => {
+    return (
+      <span {...props.attributes}>{props.children}</span>
+    );
+  },
+
+};
+export default editable;
