+++ packages/text-format/src/toolbar/index.tsx
@@ -1,0 +1,12 @@
+import { Text } from '@bento-editor/core';
+import React from 'react';
+
+const toolbar: Text['toolbar'] = {
+  Icon: () => {
+    return (
+      <div>F</div>
+    );
+  },
+
+};
+export default toolbar;
